revision: make krate immutable
