+++ packages/cli/src/cli/serve.rs
@@ -55,12 +55,12 @@ impl ServeArgs {
     }
 
     pub(crate) fn load_krate(&mut self) -> Result<DioxusCrate> {
-        let mut krate = DioxusCrate::new(&self.build_arguments.target_args)?;
-        self.resolve(&mut krate)?;
+        let krate = DioxusCrate::new(&self.build_arguments.target_args)?;
+        self.resolve(&krate)?;
         Ok(krate)
     }
 
-    pub(crate) fn resolve(&mut self, krate: &mut DioxusCrate) -> Result<()> {
+    pub(crate) fn resolve(&mut self, krate: &DioxusCrate) -> Result<()> {
         // Enable hot reload.
         if self.hot_reload.is_none() {
             self.hot_reload = Some(krate.settings.always_hot_reload.unwrap_or(true));
